Fixed all warnings. - Made some casts more explicit; - Tidied up a few constructor notation initialisers of primitive types; - Fixed uint64_t-s print(f)ed as %x, an insidious source of crash; - Fixed a myriad of int / uint, (u)int / float and 32 / 64 bit implicit casts; - Fixed assignment-as-conditional in stringutils and Font; - Fixed initialization order issues; - Fixed Missing cases in switch statements; - Fixed "not all paths return" warnings in XonEntity & c; - Fixed Hash usage errors in xr3json and Name; - Fixed FloatBuffer::AllocBuffer(); - Fixed multiple instances of using the wrong INVALID_ID in Gfx; - Fixed narrowing int - bool conversion in XR::Device::Init(); - Fixed the fullscreen config option check in Device::Init(); - Fixed fseek use for 64 bits on windows; - Fixed uninitialized local in Font::CacheGlyph(); - Fixed Shader::OnLoaded(), which wasn't checking for a valid fragment shader component; - Fixed warning-error in glLoaderGen generated profile; - Fixed missing override keywords in xr3ui (and removed virtual keywords from override declarations); - Softcoded literal constants in UIBuilder; - Added missing implementation of UIGridLayout getters; - Removed legacy Renderer code causing warning-errors until update;

+++ unittests/src/SpinlockTests.cpp
@@ -1,5 +1,5 @@
 #include "gtest/gtest.h"
-#include "XR/SpinLock.hpp"
+#include "XR/Spinlock.hpp"
 #include "XR/debug.hpp"
 #include <vector>
 #include <mutex>
@@ -40,19 +40,19 @@ namespace XR
       }
     }
 
-    void ThreadWhole(int id, int n)
+    void ThreadWhole(size_t id, size_t n)
     {
       std::unique_lock<Spinlock> lock(spinner);
-      for (int i = 0; i < n; ++i)
+      for (size_t i = 0; i < n; ++i)
       {
         ++data[id];
         ++accessCounter;
       }
     }
 
-    void ThreadPerAccess(int id, int n)
+    void ThreadPerAccess(size_t id, size_t n)
     {
-      for (int i = 0; i < n; ++i)
+      for (size_t i = 0; i < n; ++i)
       {
         {
           std::unique_lock<Spinlock> lock(spinner);
@@ -73,12 +73,12 @@ namespace XR
 
     std::vector<std::thread> threads;
     threads.reserve(data.data.size());
-    for (int i = 0; i < threads.capacity(); ++i)
+    for (size_t i = 0; i < threads.capacity(); ++i)
     {
       threads.emplace_back(&Data::ThreadWhole, &data, i, numAccesses);
     }
 
-    for (int i = 0; i < threads.capacity(); ++i)
+    for (size_t i = 0; i < threads.capacity(); ++i)
     {
       threads[i].join();
     }
@@ -93,12 +93,12 @@ namespace XR
 
     std::vector<std::thread> threads;
     threads.reserve(data.data.size());
-    for (int i = 0; i < threads.capacity(); ++i)
+    for (size_t i = 0; i < threads.capacity(); ++i)
     {
       threads.emplace_back(&Data::ThreadPerAccess, &data, i, numAccesses);
     }
 
-    for (int i = 0; i < threads.capacity(); ++i)
+    for (size_t i = 0; i < threads.capacity(); ++i)
     {
       threads[i].join();
     }
